Reorganize pages and components
