+++ client/src/pages/Home.js
@@ -1,9 +1,9 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { logoutUser } from "../../actions/authActions";
+import { logoutUser } from "../actions/authActions";
 
-class Dashboard extends Component {
+class Home extends Component {
   onLogoutClick = e => {
     e.preventDefault();
     this.props.logoutUser();
@@ -21,7 +21,7 @@ class Dashboard extends Component {
   }
 }
 
-Dashboard.propTypes = {
+Home.propTypes = {
   auth: PropTypes.object.isRequired
 };
 
@@ -33,4 +33,4 @@ const mapStateToProps = state => ({
 export default connect(
   mapStateToProps,
   { logoutUser }
-)(Dashboard);
+)(Home);
